Revert "Fix URL of Artifactory service with new floating IP"
This reverts commit e03e938e21635f078cd8e889d11cf64480669e92.

We need to revert the commit hash to the last commit before the Experiment Design service refactor. That contains some problems that we need to deal in another task.
Some later commits also contains wrong submodule and webpack dependency changes.
We are going to deal with that later.
We already have a hotfix branch to merge to this modified develop branch that contains all the useful commits that we need for now.